Access specifiers added and backend validation

+++ src/main/java/com/acms/CentralSellerPortal/Entities/Product.java
@@ -1,6 +1,7 @@
 package com.acms.CentralSellerPortal.Entities;
 
 import javax.persistence.*;
+import javax.validation.Constraint;
 import javax.validation.constraints.NotNull;
 import com.fasterxml.jackson.annotation.JsonIgnore;
 
@@ -39,7 +40,12 @@ public class Product implements Serializable {
     public Product() {
     }
 
-    public Product(long product_id, @NotNull String product_name, @NotNull int price, @NotNull String product_description, @NotNull int discount, Seller seller) {
+    public Product(int product_id,
+                   @NotNull String product_name,
+                   @NotNull int price,
+                   @NotNull String product_description,
+                   @NotNull int discount,
+                   Seller seller) {
         this.productId = product_id;
         this.productName = product_name;
         this.price = price;
